Reorganized files structure.
