+++ src/vs15/CoCo/Providers/CSharpClassifierProvider.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using CoCo.Analyser;
 using CoCo.Analyser.CSharp;
+using CoCo.Editor;
 using CoCo.Services;
 using CoCo.Settings;
 using CoCo.Utils;
@@ -24,7 +25,7 @@ namespace CoCo.Providers
         /// <summary>
         /// Determines that settings was set to avoid a many sets settings from the classifier
         /// </summary>
-        private static bool _wasSettingsSet;
+        private static bool _wereSettingsSet;
 
         private readonly Dictionary<string, ClassificationInfo> _classificationsInfo;
 
@@ -52,13 +53,13 @@ namespace CoCo.Providers
         public IClassifier GetClassifier(ITextBuffer textBuffer)
         {
             MigrationService.MigrateSettingsTo_2_0_0();
-            if (!_wasSettingsSet)
+            if (!_wereSettingsSet)
             {
                 var settings = SettingsManager.LoadEditorSettings(Paths.CoCoSettingsFile, MigrationService.Instance);
                 var option = OptionService.ToOption(settings);
                 FormattingService.SetFormattingOptions(option);
                 AnalyzingService.SetAnalyzingOptions(option);
-                _wasSettingsSet = true;
+                _wereSettingsSet = true;
             }
 
             return textBuffer.Properties.GetOrCreateSingletonProperty(() =>
