+++ src/vs15/CoCo/Providers/VisualBasicClassifierProvider.cs
@@ -2,6 +2,7 @@
 using System.ComponentModel.Composition;
 using CoCo.Analyser;
 using CoCo.Analyser.VisualBasic;
+using CoCo.Editor;
 using CoCo.Services;
 using CoCo.Settings;
 using CoCo.Utils;
@@ -12,7 +13,7 @@ using Microsoft.VisualStudio.Utilities;
 namespace CoCo.Providers
 {
     /// <summary>
-    /// Classifier provider which adds <see cref="VisualBasicClassifierProvider"/> to the set of classifiers.
+    /// Classifier provider which adds <see cref="VisualBasicClassifier"/> to the set of classifiers.
     /// </summary>
     [Export(typeof(IClassifierProvider))]
     [ContentType("Basic")]
@@ -21,7 +22,7 @@ namespace CoCo.Providers
         /// <summary>
         /// Determines that settings was set to avoid a many sets settings from the classifier
         /// </summary>
-        private static bool _wasSettingsSet;
+        private static bool _wereSettingsSet;
 
         private readonly Dictionary<string, ClassificationInfo> _classificationsInfo;
 
@@ -48,13 +49,13 @@ namespace CoCo.Providers
         public IClassifier GetClassifier(ITextBuffer textBuffer)
         {
             MigrationService.MigrateSettingsTo_2_0_0();
-            if (!_wasSettingsSet)
+            if (!_wereSettingsSet)
             {
                 var settings = SettingsManager.LoadEditorSettings(Paths.CoCoSettingsFile, MigrationService.Instance);
                 var option = OptionService.ToOption(settings);
                 FormattingService.SetFormattingOptions(option);
                 AnalyzingService.SetAnalyzingOptions(option);
-                _wasSettingsSet = true;
+                _wereSettingsSet = true;
             }
 
             return textBuffer.Properties.GetOrCreateSingletonProperty(() =>
