+++ src/vs15/CoCo/QuickInfo/QuickInfoChangingService.cs
@@ -1,28 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using CoCo.UI.Data;
 
-namespace CoCo.Services
+namespace CoCo.QuickInfo
 {
-    public enum QuickInfoState
-    {
-        Disable = 0,
-        Extend = 1,
-        Override = 2,
-    }
-
-    public delegate void QuickInfoChangedEventHandler(QuickInfoChangedEventArgs args);
-
-    public class QuickInfoChangedEventArgs : EventArgs
-    {
-        public QuickInfoChangedEventArgs(IReadOnlyDictionary<string, QuickInfoState> changedOptions)
-        {
-            Changes = changedOptions;
-        }
-
-        public IReadOnlyDictionary<string, QuickInfoState> Changes { get; }
-    }
-
     internal sealed class QuickInfoChangingService
     {
         public static readonly QuickInfoChangingService Instance = new QuickInfoChangingService();
